Build 1.0.125

+++ class_fan.py
@@ -258,7 +258,7 @@ class FanClass(QThread):
             # s = s if s < 10 else 10
             # # s = int((20 - (10 - s)) / 5) + 1  # 5 Speed
             # s = int((10 - s) / 4) + 1   # 6 Speed
-            s = ((10 - s) / 2) - 4
+            s = int((10 - s) / 2) - 4
             self.switch(s)
             print(self.id, " PID ", speed_raw, " Sw ", s)
             # if self._logging_t and self.fan_controller.master_mode == MASTER:
